Fix warnings and tests

+++ include/boost/dll/detail/elf_info.hpp
@@ -190,7 +190,7 @@ private:
                 f_.read((char*)&symbols[0], section.sh_size);
                 f_.seekg(pos);
             } else if (section.sh_type == SHT_STRTAB_) {
-                text.resize(section.sh_size);
+                text.resize(static_cast<std::size_t>(section.sh_size));
 
                 const boost::filesystem::ifstream::pos_type pos = f_.tellg();
                 f_.seekg(section.sh_offset);
